+++ R/imports.R
@@ -1,0 +1,11 @@
+#' @importFrom GGally stat_prop
+#' @export
+GGally::stat_prop
+
+#' @importFrom GGally stat_weighted_mean
+#' @export
+GGally::stat_weighted_mean
+
+#' @importFrom GGally stat_cross
+#' @export
+GGally::stat_cross