+++ NEWS.md
@@ -1,5 +1,9 @@
 # JLutils (development version)
 
+* `stat_prop()`, `stat_weigthed_mean()` and `stat_cross()` have
+  been moved to `GGally` (>= 2.0.0)
+* `stat_fill_labels()` & `stat_stack_labels()` are deprecated.
+  Use `GGally::stat_prop()` instead.
 * `ggcross()` has been updated.
 * `ggchisq_res()` is now deprecated. You could use instead `GGally::ggtable()` or `JLutils::ggcross()`.
 * `tidy_detailed()` keeps now the original order of terms.